requirements.txt
flask==2.0.3
requests==2.27.1
requests_oauthlib==1.3.1
PyYAML==6.0
peewee==3.14.8
pymysql==1.0.2
cryptography==36.0.1